@@ -40,6 +40,6 @@
         "codelyzer": "4.0.1",
         "ts-node": "4.1.0",
         "tslint": "5.9.1",
-        "typescript": "2.6.2"
+        "typescript": "^2.7.2"
     }
 }
